Project now adapts for multiple templates
Refactored the structure of the files so that the script can be used to generate more than just Makefile templates.

The primary script (template.sh) that is executed with the alias 'template' from anywhere in the system will now:
    - prompt the user to select a secondary script (with fzf) corresponding to the desired template type
    - The secondary script will execute accordingly.

As things stand, the makefile.mk script will prompt the user for the desired Makefile template through fzf as before.

Further secondary scripts can be added to the /scripts folder.

A /bin folder is no longer created in .config

All of the necessary resources for the execution of the primary script will be stored in a "templates" folder in the .config folder.